+++ physical_ai_manager/src/components/EpisodeStatus.js
@@ -17,6 +17,8 @@
 import React from 'react';
 import clsx from 'clsx';
 
+import { useSelector } from 'react-redux';
+
 const classEpisodeStatusBody = clsx(
   'h-full',
   'w-full',
@@ -37,7 +39,10 @@ const classEpisodeStatusBody = clsx(
   'bg-white'
 );
 
-export default function EpisodeStatus({ episodeStatus }) {
+export default function EpisodeStatus() {
+  const taskStatus = useSelector((state) => state.tasks.taskStatus);
+  const taskInfo = useSelector((state) => state.tasks.taskInfo);
+
   return (
     <div className={classEpisodeStatusBody}>
       <div className="mb-1 justify- text-3xl" style={{ fontSize: 'clamp(1.5rem, 1.5vw, 2rem)' }}>
@@ -48,8 +53,8 @@ export default function EpisodeStatus({ episodeStatus }) {
         className="w-full bg-gray-200 rounded-lg py-1.5 px-3 font-bold whitespace-nowrap"
         style={{ fontSize: 'clamp(1rem, 1.5vw, 2rem)' }}
       >
-        <span className="font-bold">{episodeStatus?.currentEpisodeNumber}</span> /{' '}
-        <span className="text-gray-600">{episodeStatus?.numEpisodes}</span>
+        <span className="font-bold">{taskStatus.currentEpisodeNumber}</span> /{' '}
+        <span className="text-gray-600">{taskInfo.numEpisodes}</span>
       </div>
     </div>
   );