+++ physical_ai_manager/CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package physical_ai_manager
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.5.8 (2025-07-07)
+------------------
+* Applied Redux Toolkit for better state management
+* Added heartbeat status to the UI
+* Contributors: Kiwoong Park
+
 0.5.7 (2025-06-26)
 ------------------
 * None